backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
djangorestframework-simplejwt==4.7.2
Pillow==10.0.0
django-filter==23.2.0
python-dotenv==0.21.0
urllib3==1.26.6
psycopg2-binary==2.9.7
urllib3==1.26.6
gunicorn==20.1.0
#black==23.9.1